separate Test.setup and TestLogger, because these are independent from others
